requirements.txt
aniso8601==9.0.1
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
